fix ExampleRsMeta impl
