+++ README.md
@@ -2,6 +2,15 @@
 
 GstremerをRustで書くサンプル
 
+## crate
+
+|directory|desctiption|
+|---|---|
+|meta/example-rs|カスタムメタデータ `ExampleRsMeta` のRust実装|
+|meta/example-rs-sys|カスタムメタデータ `ExampleRsMeta` をRustから操作するbinding|
+|plugin/src/testtrans|BaseTransformを使った実装|
+|plugin/src/metatrans|`ExampleRsMeta`の追加、表示などを行う実装|
+
 
 ## References
